update: tables infra, add: /records data limit
