+++ src/app/dashboard/components/TrackWorldRecords.tsx
@@ -6,9 +6,7 @@ import Link from "next/link"
 import useSWR from "swr"
 import Table, { TableOptions } from "~/ui/Table"
 import { Button } from "~/ui/Button"
-import TrackRecordsTable, {
-  trackRecordsColumns,
-} from "~/components/tables/records/TrackRecordsTable"
+import { trackRecordsColumns, trackRecordsData } from "~/components/tables/data/trackRecords"
 import { ChevronsRight } from "lucide-react"
 
 interface Props extends TableOptions {
@@ -37,7 +35,7 @@ export default function TrackWorldRecords({ tracks, ...rest }: Props) {
 
     if (isLoading) return <SkeletonTable />
 
-    return <TrackRecordsTable records={data.records} columns={trackRecordsColumns} {...rest} />
+    return <Table data={trackRecordsData(data.records)} columns={trackRecordsColumns} {...rest} />
   }
 
   return (
