+++ src/app/records/[top]/components/DynamicTableRenderer.tsx
@@ -2,127 +2,154 @@
 
 import { useEffect, useState } from "react"
 import useSwr from "swr"
+import useSWRMutation from "swr/mutation"
 import Pill from "~/components/pills/Pill"
 import Spinner from "~/components/Spinner"
-import RecordHoldersTable, {
-  recordHoldersColumnWithControls,
-} from "~/components/tables/records/RecordHoldersTable"
-import RiderWorldRecordsTableRow from "~/components/tables/expandable/RiderWorldRecordsTableRow"
-import MMRRecordsTable, {
+import Table, { TableOptions } from "~/ui/Table"
+import {
+  recordHoldersData,
+  recordHoldersColumnsWithControls,
+} from "~/components/tables/data/topRecordsHolders"
+import {
   mmrRecordsColumnsWithControls,
-} from "~/components/tables/records/MMRRecordsTable"
-import RiderRecentRacesTableRow from "~/components/tables/expandable/RiderRecentRacesTableRow"
-import SRRecordsTable, {
-  srRecordsColumnsWithControls,
-} from "~/components/tables/records/SRRecordsTable"
-import RiderSafetyStatsRow from "~/components/tables/expandable/RiderSafetyStatsRow"
-import BikeRecordsTable, { bikeRecordsColumns } from "~/components/tables/records/BikeRecordsTable"
-import ContactRecordsTable, {
+  mmrRecordsData,
+} from "~/components/tables/data/topMmrRecords"
+import { srRecordsColumnsWithControls, srRecordsData } from "~/components/tables/data/topSrRecords"
+import {
   contactRecordsColumnsWithRatio,
-} from "~/components/tables/records/ContactRecordsTable"
+  contactRecordsData,
+} from "~/components/tables/data/topContactRecords"
+import {
+  bikeRecordsColumnsWithControls,
+  bikeRecordsData,
+} from "~/components/tables/data/topBikeRecords"
+import RiderWorldRecordsTableRow from "~/components/tables/expandable/RiderWorldRecordsTableRow"
+import RiderRecentRacesTableRow from "~/components/tables/expandable/RiderRecentRacesTableRow"
+import RiderSafetyStatsRow from "~/components/tables/expandable/RiderSafetyStatsRow"
+import { fetcher } from "~/api/fetcher"
+
+const PAGE_SIZE = 20
+const LIMIT = 100
 
 export default function DynamicTableRenderer({ top, records }) {
-  return dynamicDataMap[top].render(records)
-}
+  const [limit, setLimit] = useState(LIMIT)
 
-const tableProps = {
-  defaultPageSize: 20,
-  pageSizeEnabled: true,
-  sortingEnabled: true,
-  paginationEnabled: true,
-}
+  const tableProps: TableOptions = {
+    defaultPageSize: PAGE_SIZE,
+    defaultDataCap: limit,
+    pageSizeEnabled: true,
+    paginationEnabled: true,
+    dataCapEnabled: true,
+    onDataCapChange: setLimit,
+  }
 
-const dynamicDataMap = {
-  riders: {
-    render: (records) => {
-      return (
-        <RecordHoldersTable
-          riders={records.riders}
-          columns={recordHoldersColumnWithControls}
-          sortingKeys={["records"]}
-          expandable={{
-            render: (row) => <RiderWorldRecordsTableRow row={row} />,
-          }}
-          {...tableProps}
-        />
-      )
-    },
-  },
-  mmr: {
-    render: (records) => {
-      return (
-        <MMRRecordsTable
-          riders={records.riders}
-          columns={mmrRecordsColumnsWithControls}
-          sortingKeys={["rating"]}
-          expandable={{
-            render: (row) => <RiderRecentRacesTableRow row={row} />,
-          }}
-          {...tableProps}
-        />
-      )
-    },
-  },
-  sr: {
-    render: (records) => {
-      return (
-        <SRRecordsTable
-          riders={records.riders}
-          columns={srRecordsColumnsWithControls}
-          sortingKeys={["rating", "ratio"]}
-          expandable={{
-            render: (row) => <RiderSafetyStatsRow row={row} />,
-          }}
-          {...tableProps}
-        />
-      )
-    },
-  },
-  contacts: {
-    render: (records) => {
-      return (
-        <ContactRecordsTable
-          worldContacts={records}
-          columns={contactRecordsColumnsWithRatio}
-          sortingKeys={["contacts", "ratio"]}
-          {...tableProps}
-        />
-      )
-    },
-  },
-  bikes: {
-    render: (records) => {
-      const Content = () => {
-        const { data: apiData, isLoading } = useSwr("/summary")
-        const [laps, setLaps] = useState(0)
+  const { data: topRecords, trigger, isMutating } = useSWRMutation(`/top/${top}/${limit}`, fetcher)
 
-        useEffect(() => {
-          if (apiData) setLaps(apiData.laps)
-        }, [isLoading])
+  useEffect(() => {
+    if (limit === LIMIT) return
+    trigger()
+  }, [limit])
 
+  const dynamicDataMap = {
+    riders: {
+      render: (records) => {
         return (
-          <>
-            <div className="mb-4 flex justify-end gap-2">
-              <div className="text-accent">Total Laps</div>
-              {laps ? (
-                <Pill text={laps.toLocaleString()} color="primary" />
-              ) : (
-                <div className="min-h-[28px]">
-                  <Spinner />
-                </div>
-              )}
-            </div>
-            <BikeRecordsTable
-              bikes={records.bikes}
-              columns={bikeRecordsColumns}
-              totalLaps={laps}
-              {...tableProps}
-            />
-          </>
+          <Table
+            data={recordHoldersData(records.riders)}
+            columns={recordHoldersColumnsWithControls}
+            sortingKeys={["records"]}
+            expandable={{
+              render: (row) => <RiderWorldRecordsTableRow row={row} />,
+            }}
+            {...tableProps}
+          />
         )
-      }
-
-      return <Content />
+      },
     },
-  },
+    mmr: {
+      render: (records) => {
+        return (
+          <Table
+            data={mmrRecordsData(records.riders)}
+            columns={mmrRecordsColumnsWithControls}
+            sortingKeys={["rating"]}
+            expandable={{
+              render: (row) => <RiderRecentRacesTableRow row={row} />,
+            }}
+            {...tableProps}
+          />
+        )
+      },
+    },
+    sr: {
+      render: (records) => {
+        return (
+          <Table
+            data={srRecordsData(records.riders)}
+            columns={srRecordsColumnsWithControls}
+            sortingKeys={["rating", "ratio"]}
+            expandable={{
+              render: (row) => <RiderSafetyStatsRow row={row} />,
+            }}
+            {...tableProps}
+          />
+        )
+      },
+    },
+    contacts: {
+      render: (records) => {
+        return (
+          <Table
+            data={contactRecordsData(records)}
+            columns={contactRecordsColumnsWithRatio}
+            sortingKeys={["contacts", "ratio"]}
+            expandable={{
+              render: (row) => <RiderWorldRecordsTableRow row={row} />,
+            }}
+            {...tableProps}
+          />
+        )
+      },
+    },
+    bikes: {
+      render: (records) => {
+        const Content = () => {
+          const { data: apiData, isLoading } = useSwr("/summary")
+          const [laps, setLaps] = useState(0)
+
+          useEffect(() => {
+            if (apiData) setLaps(apiData.laps)
+          }, [isLoading])
+
+          return (
+            <>
+              <div className="mb-4 flex justify-end gap-2">
+                <div className="text-accent">Total Laps</div>
+                {laps ? (
+                  <Pill text={laps.toLocaleString()} color="primary" />
+                ) : (
+                  <div className="min-h-[28px]">
+                    <Spinner />
+                  </div>
+                )}
+              </div>
+              <Table
+                data={bikeRecordsData(records.bikes, laps)}
+                columns={bikeRecordsColumnsWithControls}
+                {...tableProps}
+              />
+            </>
+          )
+        }
+
+        return <Content />
+      },
+    },
+  }
+
+  if (isMutating) return <Spinner />
+
+  if (!topRecords || isMutating) return dynamicDataMap[top].render(records)
+
+  return dynamicDataMap[top].render(topRecords)
 }
