+++ src/app/records/[top]/page.tsx
@@ -13,7 +13,7 @@ export async function generateMetadata({ params: { top } }) {
 }
 
 export default async function Page({ params: { top } }) {
-  const topRecords = await GetDynamicTopRecords(top, 1000)
+  const topRecords = await GetDynamicTopRecords(top, 100)
 
   if (!dynamicTitleMap[top]) return <Result title="Not Found" description="No records found" />
