+++ src/app/servers/page.tsx
@@ -17,7 +17,6 @@ export default async function Page() {
     <PageLayout
       width="app"
       header={{
-        backEnabled: true,
         title: "MXB Servers",
       }}
     >
