+++ src/components/tables/data/topBikeRecords.tsx
@@ -1,39 +1,20 @@
-"use client"
-
-import useSwr from "swr"
 import Spinner from "~/components/Spinner"
 import BikeTicTac from "~/components/pills/BikeTicTac"
 import Pill from "~/components/pills/Pill"
-import Table, { TableOptions } from "~/ui/Table"
 
-interface Props extends TableOptions {
-  bikes: any
-  columns: any
-  totalLaps: number
-}
-
-export default function BikeRecordsTable({ bikes, columns, totalLaps, ...rest }: Props) {
-  const data = bikes.map((bike) => {
+export const bikeRecordsData = (bikes, totalLaps = 0) =>
+  bikes.map((bike) => {
     const ratio = totalLaps ? (bike.laps / totalLaps) * 100 : 0
-    return {
+    const res = {
       _id: bike.name + bike.laps,
       name: bike.name,
       laps: bike.laps,
       ratio: ratio,
     }
+    return res
   })
-  // console.log("%cBikeRecordsTable", "color: steelblue", { bikes: data })
 
-  const sortKeys = ["laps", "ratio"]
-
-  return (
-    <div className="flex flex-col items-end">
-      <Table data={data} columns={columns} sortingKeys={sortKeys} {...rest} />
-    </div>
-  )
-}
-
-export const bikeRecordsColumns = [
+export const bikeRecordsColumnsWithControls = [
   {
     key: "name",
     label: "Bike",
