+++ src/components/tables/data/topSrRecords.tsx
@@ -1,29 +1,19 @@
-"use client"
-
 import RiderLink from "~/components/RiderLink"
-import Table, { TableOptions } from "~/ui/Table"
+import Pill from "~/components/pills/Pill"
+import { handleHPLColor } from "./topContactRecords"
 
-interface Props extends TableOptions {
-  riders: any
-  columns: any
-}
+export const srRecordsData = (records) =>
+  records.map((r) => {
+    const laps = Object.keys(r.bikes).reduce((acc, curr) => acc + r.bikes[curr].laps, 0)
+    return {
+      _id: r._id,
+      name: r.name,
+      rating: r.SR,
+      ratio: Math.ceil((r.contact / laps) * 100) / 100,
+    }
+  })
 
-export default function MMRRecordsTable({ riders, columns, ...rest }: Props) {
-  const data: any = riders.map((r) => ({
-    _id: r._id,
-    name: r.name,
-    rating: r.MMR,
-  }))
-  // console.log("%cMMRRecordsTable", "color: steelblue", { riders: data })
-
-  return (
-    <div className="flex flex-col items-end">
-      <Table data={data} columns={columns} {...rest} />
-    </div>
-  )
-}
-
-export const mmrRecordsColumns = [
+export const srRecordsColumns = [
   {
     key: "name",
     label: "Rider",
@@ -38,7 +28,7 @@ export const mmrRecordsColumns = [
   },
 ]
 
-export const mmrRecordsColumnsWithControls = [
+export const srRecordsColumnsWithControls = [
   {
     key: "name",
     label: "Rider",
@@ -51,5 +41,10 @@ export const mmrRecordsColumnsWithControls = [
     key: "rating",
     label: "Rating",
     align: "right",
+  },
+  {
+    key: "ratio",
+    label: "Hits per lap",
+    render: (ratio) => <Pill color={handleHPLColor(ratio)} text={ratio.toFixed(2)} />,
   },
 ]
