+++ src/components/tables/expandable/RiderRecentRacesTableRow.tsx
@@ -2,8 +2,9 @@
 
 import useSWR from "swr"
 import Spinner from "~/components/Spinner"
+import Table from "~/ui/Table"
+import { riderRecentRacesColumns, riderRecentRacesData } from "../data/riderRecentRaces"
 import { handleRacismSanitization } from "~/utils/handleRacismSanitization"
-import RiderRecentRacesTable, { riderRecentRacesColumns } from "../rider/RiderRecentRacesTable"
 
 export default function RiderRecentRacesTableRow({ row }) {
   const { data: raceData, isLoading } = useSWR(`/rider/${row._id}/races`)
@@ -20,7 +21,7 @@ export default function RiderRecentRacesTableRow({ row }) {
       <div className="mb-2 text-lg font-semibold">
         {handleRacismSanitization(row.name)}&apos;s Last Ten Races
       </div>
-      <RiderRecentRacesTable races={raceData.races} columns={riderRecentRacesColumns} />
+      <Table data={riderRecentRacesData(raceData.races)} columns={riderRecentRacesColumns} />
     </>
   )
 }
